+++ lib/mina_sidekiq.rb
@@ -1,4 +1,0 @@
-require 'mina_sidekiq/version'
-
-module MinaSidekiq
-end