+++ lib/mina_sidekiq_upstart.rb
@@ -1,0 +1,4 @@
+require 'mina_sidekiq_upstart/version'
+
+module MinaSidekiqUpstart
+end